fix(a11y): moved role and aria-checked attributes to the input element
