+++ projects/angular/clarity.api.md
@@ -764,6 +764,8 @@ export class ClrCalendar implements OnDestroy {
 export class ClrCheckbox extends WrappedFormControl<ClrCheckboxWrapper> {
     constructor(vcr: ViewContainerRef, injector: Injector, control: NgControl, renderer: Renderer2, el: ElementRef<HTMLInputElement>, toggle: string);
     // (undocumented)
+    protected get controlChecked(): boolean;
+    // (undocumented)
     get controlDisabled(): boolean;
     // (undocumented)
     ngOnInit(): void;
@@ -819,11 +821,9 @@ export class ClrCheckboxModule {
 
 // @public (undocumented)
 export class ClrCheckboxWrapper implements OnInit, OnDestroy {
-    constructor(toggleService: BehaviorSubject<boolean>, el: ElementRef<HTMLElement>);
+    constructor(toggleService: BehaviorSubject<boolean>);
     // (undocumented)
     checkbox: ClrCheckbox;
-    // (undocumented)
-    protected get controlChecked(): boolean;
     // (undocumented)
     label: ClrLabel;
     // (undocumented)
