+++ projects/angular/src/forms/checkbox/checkbox-wrapper.ts
@@ -5,16 +5,7 @@
  * The full license information can be found in LICENSE in the root directory of this project.
  */
 
-import {
-  Component,
-  ContentChild,
-  ElementRef,
-  forwardRef,
-  Inject,
-  InjectionToken,
-  OnDestroy,
-  OnInit,
-} from '@angular/core';
+import { Component, ContentChild, forwardRef, Inject, InjectionToken, OnDestroy, OnInit } from '@angular/core';
 import { BehaviorSubject, Subscription } from 'rxjs';
 
 import { ClrLabel } from '../common/label';
@@ -38,8 +29,6 @@ export const IS_TOGGLE_PROVIDER = { provide: IS_TOGGLE, useFactory: isToggleFact
     '[class.clr-checkbox-wrapper]': '!toggle',
     '[class.clr-checkbox-wrapper-disabled]': 'checkbox?.controlDisabled',
     '[class.clr-toggle-wrapper]': 'toggle',
-    '[attr.role]': 'toggle ? "switch" : null',
-    '[attr.aria-checked]': 'toggle ? controlChecked : null',
   },
   providers: [ControlIdService, IS_TOGGLE_PROVIDER],
 })
@@ -49,16 +38,12 @@ export class ClrCheckboxWrapper implements OnInit, OnDestroy {
   toggle = false;
   private subscriptions: Subscription[] = [];
 
-  constructor(@Inject(IS_TOGGLE) toggleService: BehaviorSubject<boolean>, private el: ElementRef<HTMLElement>) {
+  constructor(@Inject(IS_TOGGLE) toggleService: BehaviorSubject<boolean>) {
     this.subscriptions.push(
       toggleService.subscribe(state => {
         this.toggle = state;
       })
     );
-  }
-
-  protected get controlChecked() {
-    return this.el.nativeElement?.querySelector<HTMLInputElement>("input[type='checkbox']")?.checked;
   }
 
   ngOnInit() {
